azure/azure-run-arm-py/requirements.txt
python-dotenv==0.15.0
azure-mgmt-resource==15.0.0
azure-identity==1.5.0
msrestazure==0.6.4